Adding project files for RAD Studio XE6.
Adding TSelectionEditor classes to design-time packages for a few components (more to follow in the future) to add dependent units to 'uses' clauses.

Misc tweaks to TIdIOHandler for timeout handling.

Signature change for ParseMetaHTTPEquiv().

Updated SASL components to support Initial Client Responses.

Updated BytesToString() to not make a copy of the input TIdBytes if the entire array is to be converted.

Added IndexOfIP() method to TIdStackLocalAddressList.

Added GetLocalAddressList() method to TIdStack and deprecated AddLocalAddressesToList() method.  Updated TIdIPAddrMon to use TIdStack.GetLocalAddressList().

git-svn-id: https://svn.atozed.com:444/svn/Indy10/trunk@5105 421b718a-868e-4740-9b41-4026df17543c