Adding P(U)IntX type declarations to IdGlobal.pas for Delphi.
Updating TIdMessageClient.ReceiveBody() to force the message body to be read as 8bit when TIdMessage.NoDecode is true, instead of using the TIdMessage.ContentTransferEncoding property.

Updating TIdMessageDecoderMIME.ReadBody() to prepare to handle different line breaks (CRLF vs bare-LF) when reading in binary mode.  Not actually implemented at this time, CRLF is still expected for now until a detection method can be implemented.

git-svn-id: https://svn.atozed.com:444/svn/Indy10/trunk@5285 421b718a-868e-4740-9b41-4026df17543c